+++ R/archive.R
@@ -1,7 +1,7 @@
 
-renv_archive_type <- function(path) {
+renv_archive_type <- function(archive) {
 
-  ext <- fileext(path)
+  ext <- fileext(archive)
 
   if (ext %in% c(".tgz", ".tar", ".tar.gz"))
     return("tar")
@@ -12,98 +12,58 @@ renv_archive_type <- function(path) {
 
 }
 
-renv_archive_list <- function(path) {
-  suppressWarnings(renv_archive_list_impl(path))
+renv_archive_list <- function(archive) {
+  suppressWarnings(renv_archive_list_impl(archive))
 }
 
-renv_archive_list_impl <- function(path) {
+renv_archive_list_impl <- function(archive) {
 
   switch(
-    renv_archive_type(path),
-    tar = untar(path, list = TRUE),
-    zip = unzip(path, list = TRUE)[["Name"]],
-    stopf("don't know how to list files in archive '%s'", basename(path))
+    renv_archive_type(archive),
+    tar = untar(archive, list = TRUE),
+    zip = unzip(archive, list = TRUE)[["Name"]],
+    stopf("don't know how to list files in archive '%s'", basename(archive))
   )
 
 }
 
-renv_archive_decompress <- function(path, exdir = ".", ...) {
+renv_archive_decompress <- function(archive, files = NULL, exdir = ".", ...) {
 
   switch(
-    renv_archive_type(path),
-    tar = renv_archive_decompress_tar(path, exdir = exdir, ...),
-    zip = renv_archive_decompress_zip(path, exdir = exdir, ...),
-    stopf("don't know how to decompress archive '%s'", basename(path))
+    renv_archive_type(archive),
+    tar = renv_archive_decompress_tar(archive, files = files, exdir = exdir, ...),
+    zip = renv_archive_decompress_zip(archive, files = files, exdir = exdir, ...),
+    stopf("don't know how to decompress archive '%s'", basename(archive))
   )
 
 }
 
-renv_archive_decompress_tar_find <- function() {
+renv_archive_decompress_tar <- function(archive, files = NULL, exdir = ".", ...) {
 
-  # check for tar in envvar -- allow for explicitly-requested
-  # internal tar as well
-  tar <- Sys.getenv("TAR", unset = NA)
-  if (identical(tar, "internal"))
-    return(NULL)
-
-  # if the requested tar exists, use it
-  if (!is.na(tar) && nzchar(Sys.which(tar)))
-    return(tar)
-
-  # no TAR envvar set; try looking for tar on the PATH
-  # TODO: is this safe on Windows? what if a bad tar is on the PATH?
-  tar <- Sys.which("tar")
+  # if an appropriate system tar is available, use it
+  tar <- renv_tar_exe()
   if (nzchar(tar))
-    return(tar)
-
-  # no tar found
-  NULL
-
-}
-
-renv_archive_decompress_tar <- function(path, exdir = ".", ...) {
-
+    return(renv_tar_decompress(tar, archive = archive, files = files, exdir = exdir, ...))
 
   # when using internal TAR, we want to suppress warnings
   # (otherwise we get noise about global PAX headers)
-  tar <- renv_archive_decompress_tar_find()
-  if (is.null(tar)) {
-    suppressWarnings(untar(path, exdir = exdir, tar = "internal", ...))
-    return(TRUE)
-  }
-
-  # TODO: is it safe to use an external tar on Windows?
-  # should we validate that the version of tar.exe found
-  # on the PATH is okay?
-
-  # construct arguments for archive extraction
-  path <- renv_path_normalize(path, winslash = "/", mustWork = TRUE)
-  args <- c("xf", shQuote(path))
-
-  # add in extraction dir
-  if (exdir != ".") {
-    ensure_directory(exdir)
-    exargs <- c("-C", shQuote(renv_path_normalize(exdir, winslash = "/", mustWork = TRUE)))
-    args <- c(args, exargs)
-  }
-
-  # execute the command
-  status <- system2(tar, args, stdout = FALSE, stderr = FALSE)
-  if (status != 0L)
-    stopf("error decompressing '%s' [error code %i]", basename(path), status)
-
-  TRUE
+  suppressWarnings(untar(archive, files = files, exdir = exdir, tar = "internal", ...))
+  return(TRUE)
 
 }
 
-renv_archive_decompress_zip <- function(path, ...) {
+renv_archive_decompress_zip <- function(archive, files = NULL, exdir = ".", ...) {
 
   # the default unzip tool will give warnings rather than
   # errors if R was unable to extract from a zip archive
-  status <- tryCatch(unzip(path, ...), condition = identity)
+  status <- tryCatch(
+    unzip(archive, files = files, exdir = exdir, ...),
+    condition = identity
+  )
+
   if (inherits(status, "condition")) {
     fmt <- "failed to decompress '%s' [%s]"
-    message <- sprintf(fmt, basename(path), conditionMessage(status))
+    message <- sprintf(fmt, basename(archive), conditionMessage(status))
     stop(simpleError(message))
   }
 
@@ -111,3 +71,47 @@ renv_archive_decompress_zip <- function(path, ...) {
 
 }
 
+renv_archive_find <- function(archive, pattern) {
+  files <- renv_archive_list(archive)
+  grep(pattern, files, value = TRUE)
+}
+
+renv_archive_read <- function(archive, file) {
+
+  type <- renv_archive_type(archive)
+  case(
+    type == "tar" ~ renv_archive_read_tar(archive, file),
+    type == "zip" ~ renv_archive_read_zip(archive, file),
+    ~ stopf("don't know how to read file from archive %s", renv_path_pretty(archive))
+  )
+
+}
+
+renv_archive_read_tar <- function(archive, file) {
+
+  # if an appropriate tar is available, use it
+  tar <- renv_tar_exe()
+  if (nzchar(tar)) {
+    args <- c("xf", shQuote(archive), "-O", shQuote(file))
+    return(renv_system_exec(tar, args, action = "reading file from archive"))
+  }
+
+  # create extraction directory
+  exdir <- renv_scope_tempfile("renv-archive-")
+  ensure_directory(exdir)
+
+  # unpack the requested file
+  suppressWarnings(untar(archive, files = file, exdir = exdir, tar = "internal"))
+
+  # and read it
+  archive <- file.path(exdir, file)
+  readLines(archive, warn = FALSE)
+
+}
+
+renv_archive_read_zip <- function(archive, file) {
+  renv_scope_tempdir()
+  conn <- unz(archive, file, encoding = "native.enc")
+  on.exit(close(conn), add = TRUE)
+  readLines(conn, warn = FALSE)
+}