+++ R/bootstrap.R
@@ -88,16 +88,20 @@ renv_bootstrap_download <- function(version) {
   nv <- numeric_version(version)
   components <- unclass(nv)[[1]]
 
-  methods <- if (length(components) == 4L) {
-    list(
+  # if this appears to be a development version of 'renv', we'll
+  # try to restore from github
+  dev <- length(components) == 4L
+
+  # begin collecting different methods for finding renv
+  methods <- c(
+    renv_bootstrap_download_tarball,
+    if (dev)
       renv_bootstrap_download_github
-    )
-  } else {
-    list(
+    else c(
       renv_bootstrap_download_cran_latest,
       renv_bootstrap_download_cran_archive
     )
-  }
+  )
 
   for (method in methods) {
     path <- tryCatch(method(version), error = identity)
@@ -231,6 +235,33 @@ renv_bootstrap_download_cran_archive <- function(version) {
 
   message("FAILED")
   return(FALSE)
+
+}
+
+renv_bootstrap_download_tarball <- function(version) {
+
+  # if the user has provided the path to a tarball via
+  # an environment variable, then use it
+  tarball <- Sys.getenv("RENV_BOOTSTRAP_TARBALL", unset = NA)
+  if (is.na(tarball))
+    return()
+
+  # allow directories
+  info <- file.info(tarball, extra_cols = FALSE)
+  if (identical(info$isdir, TRUE)) {
+    name <- sprintf("renv_%s.tar.gz", version)
+    tarball <- file.path(tarball, name)
+  }
+
+  # bail if it doesn't exist
+  if (!file.exists(tarball))
+    return()
+
+  fmt <- "* Bootstrapping with tarball at path '%s'."
+  msg <- sprintf(fmt, tarball)
+  message(msg)
+
+  tarball
 
 }
 
@@ -661,12 +692,16 @@ renv_bootstrap_project_type <- function(path) {
 
 }
 
-renv_bootstrap_user_dir <- function(path) {
-  dir <- renv_bootstrap_user_dir_impl(path)
-  chartr("\\", "/", dir)
+renv_bootstrap_user_dir <- function() {
+  chartr("\\", "/", renv_bootstrap_user_dir_impl())
 }
 
-renv_bootstrap_user_dir_impl <- function(path) {
+renv_bootstrap_user_dir_impl <- function() {
+
+  # use local override if set
+  override <- getOption("renv.userdir.override")
+  if (!is.null(override))
+    return(override)
 
   # use R_user_dir if available
   tools <- asNamespace("tools")
@@ -677,10 +712,8 @@ renv_bootstrap_user_dir_impl <- function(path) {
   envvars <- c("R_USER_CACHE_DIR", "XDG_CACHE_HOME")
   for (envvar in envvars) {
     root <- Sys.getenv(envvar, unset = NA)
-    if (!is.na(root)) {
-      path <- file.path(root, "R/renv")
-      return(path)
-    }
+    if (!is.na(root))
+      return(file.path(root, "R/renv"))
   }
 
   # use platform-specific default fallbacks
@@ -692,3 +725,4 @@ renv_bootstrap_user_dir_impl <- function(path) {
     "~/.cache/R/renv"
 
 }
+