+++ R/zzz.R
@@ -2,14 +2,24 @@
 .onLoad <- function(libname, pkgname) {
 
   renv_platform_init()
-  renv_methods_init()
-  renv_patch_init()
   renv_envvars_init()
-  renv_libpaths_init()
+  renv_log_init()
+  renv_methods_init()
   renv_filebacked_init()
+  renv_libpaths_init()
+  renv_patch_init()
 
   addTaskCallback(renv_repos_init_callback)
   addTaskCallback(renv_snapshot_auto_callback)
+
+  # if an renv project already appears to be loaded, then re-activate
+  # the sandbox now -- this is primarily done to support suspend and
+  # resume with RStudio where the user profile might not be run
+  if (renv_rstudio_available()) {
+    project <- getOption("renv.project.path")
+    if (!is.null(project))
+      renv_sandbox_activate(project = project)
+  }
 
 }
 
@@ -20,23 +30,31 @@
 
 renv_zzz_run <- function() {
 
-  # only run when devtools::document() is called
-  ok <- FALSE
+  # check if we're running devtools::document()
+  documenting <- FALSE
   document <- parse(text = "devtools::document")[[1]]
   for (call in sys.calls()) {
     if (identical(call[[1]], document)) {
-      ok <- TRUE
+      documenting <- TRUE
       break
     }
   }
 
-  if (!ok)
-    return(FALSE)
+  # if so, then create some files
+  if (documenting) {
+    renv_zzz_bootstrap()
+    renv_zzz_docs()
+  }
 
-  renv_zzz_bootstrap()
-  renv_zzz_docs()
+  # check if we're running as part of R CMD build
+  # if so, build our local repository with a copy of ourselves
+  building <-
+    !is.na(Sys.getenv("R_CMD", unset = NA)) &&
+    grepl("Rbuild", basename(dirname(getwd())))
 
-  TRUE
+  if (building) {
+    renv_zzz_repos()
+  }
 
 }
 
@@ -73,7 +91,11 @@ renv_zzz_docs <- function() {
 
     ensure_directory("inst/doc")
 
-    files <- list.files("vignettes")
+    files <- list.files(
+      path = "vignettes",
+      pattern = "[.](?:R|Rmd|html)$",
+    )
+
     src <- file.path("vignettes", files)
     tgt <- file.path("inst/doc", files)
     file.copy(src, tgt, overwrite = TRUE)
@@ -81,6 +103,47 @@ renv_zzz_docs <- function() {
     writef("Done!")
 
   }, onexit = TRUE)
+
+}
+
+renv_zzz_repos <- function() {
+
+  # don't run if we're running tests
+  if (renv_package_checking())
+    return()
+
+  # prevent recursion
+  installing <- Sys.getenv("RENV_INSTALLING_REPOS", unset = NA)
+  if (!is.na(installing))
+    return()
+
+  Sys.setenv("RENV_INSTALLING_REPOS" = "TRUE")
+
+  writeLines("** installing renv to package-local repository")
+
+  # get package directory
+  pkgdir <- getwd()
+
+  # move to build directory
+  tdir <- tempfile("renv-build-")
+  ensure_directory(tdir)
+  owd <- setwd(tdir)
+  on.exit(setwd(owd), add = TRUE)
+
+  # build renv again
+  r_cmd_build("renv", path = pkgdir)
+
+  # copy built tarball to inst folder
+  src <- list.files(tdir, full.names = TRUE)
+  tgt <- file.path(pkgdir, "inst/repos/src/contrib")
+
+  ensure_directory(tgt)
+  file.copy(src, tgt)
+
+  # write PACKAGES
+  renv_scope_envvars(R_DEFAULT_SERIALIZE_VERSION = "2")
+  tools::write_PACKAGES(tgt, type = "source")
+
 }
 
 if (identical(.packageName, "renv"))